move password model to user file
